+++ example/config.js
@@ -5,33 +5,33 @@ export default [
     list: [
       {
         path: '/button',
-        title: 'Button 按钮'
+        title: 'Button 按钮',
       },
       {
         path: '/cell',
-        title: 'Cell 单元格'
+        title: 'Cell 单元格',
       },
       {
         path: '/icon',
-        title: 'Icon 图标'
+        title: 'Icon 图标',
       },
       {
         path: '/image',
-        title: 'Image 图片'
+        title: 'Image 图片',
       },
       {
         path: '/col',
-        title: 'Layout 布局'
+        title: 'Layout 布局',
       },
       {
         path: '/popup',
-        title: 'Popup 弹出层'
+        title: 'Popup 弹出层',
       },
       {
         path: '/transition',
-        title: 'Transition 动画'
-      }
-    ]
+        title: 'Transition 动画',
+      },
+    ],
   },
   {
     groupName: '表单组件',
@@ -43,49 +43,49 @@ export default [
       },
       {
         path: '/checkbox',
-        title: 'Checkbox 复选框'
+        title: 'Checkbox 复选框',
       },
       {
         path: '/datetime-picker',
-        title: 'DatetimePicker 时间选择'
+        title: 'DatetimePicker 时间选择',
       },
       {
         path: '/field',
-        title: 'Field 输入框'
+        title: 'Field 输入框',
       },
       {
         path: '/picker',
-        title: 'Picker 选择器'
+        title: 'Picker 选择器',
       },
       {
         path: '/radio',
-        title: 'Radio 单选框'
+        title: 'Radio 单选框',
       },
       {
         path: '/rate',
-        title: 'Rate 评分'
+        title: 'Rate 评分',
       },
       {
         path: '/search',
-        title: 'Search 搜索'
+        title: 'Search 搜索',
       },
       {
         path: '/slider',
-        title: 'Slider 滑块'
+        title: 'Slider 滑块',
       },
       {
         path: '/stepper',
-        title: 'Stepper 步进器'
+        title: 'Stepper 步进器',
       },
       {
         path: '/switch',
-        title: 'Switch 开关'
+        title: 'Switch 开关',
       },
       {
         path: '/uploader',
-        title: 'Uploader 文件上传'
-      }
-    ]
+        title: 'Uploader 文件上传',
+      },
+    ],
   },
   {
     groupName: '反馈组件',
@@ -93,37 +93,41 @@ export default [
     list: [
       {
         path: '/action-sheet',
-        title: 'ActionSheet 上拉菜单'
+        title: 'ActionSheet 上拉菜单',
       },
       {
         path: '/dialog',
-        title: 'Dialog 弹出框'
+        title: 'Dialog 弹出框',
       },
       {
         path: '/dropdown-menu',
-        title: 'DropdownMenu 下拉菜单'
+        title: 'DropdownMenu 下拉菜单',
       },
       {
         path: '/loading',
-        title: 'Loading 加载'
+        title: 'Loading 加载',
       },
       {
         path: '/notify',
-        title: 'Notify 消息通知'
+        title: 'Notify 消息通知',
       },
       {
         path: '/overlay',
-        title: 'Overlay 遮罩层'
+        title: 'Overlay 遮罩层',
+      },
+      {
+        path: '/share-sheet',
+        title: 'ShareSheet 分享面板',
       },
       {
         path: '/swipe-cell',
-        title: 'SwipeCell 滑动单元格'
+        title: 'SwipeCell 滑动单元格',
       },
       {
         path: '/toast',
-        title: 'Toast 轻提示'
-      }
-    ]
+        title: 'Toast 轻提示',
+      },
+    ],
   },
   {
     groupName: '展示组件',
@@ -131,53 +135,57 @@ export default [
     list: [
       {
         path: '/circle',
-        title: 'Circle 进度条'
+        title: 'Circle 进度条',
       },
       {
         path: '/collapse',
-        title: 'Collapse 折叠面板'
+        title: 'Collapse 折叠面板',
       },
       {
         path: '/count-down',
-        title: 'CountDown 倒计时'
+        title: 'CountDown 倒计时',
       },
       {
         path: '/divider',
-        title: 'Divider 分割线'
+        title: 'Divider 分割线',
+      },
+      {
+        path: '/empty',
+        title: 'Empty 空状态',
       },
       {
         path: '/notice-bar',
-        title: 'NoticeBar 通告栏'
+        title: 'NoticeBar 通告栏',
       },
       {
         path: '/panel',
-        title: 'Panel 面板'
+        title: 'Panel 面板',
       },
       {
         path: '/progress',
-        title: 'Progress 进度条'
+        title: 'Progress 进度条',
       },
       {
         path: '/skeleton',
-        title: 'Skeleton 骨架屏'
+        title: 'Skeleton 骨架屏',
       },
       {
         path: '/steps',
-        title: 'Steps 步骤条'
+        title: 'Steps 步骤条',
       },
       {
         path: '/sticky',
-        title: 'Sticky 粘性布局'
+        title: 'Sticky 粘性布局',
       },
       {
         path: '/tag',
-        title: 'Tag 标记'
+        title: 'Tag 标记',
       },
       {
         path: '/tree-select',
-        title: 'TreeSelect 分类选择'
-      }
-    ]
+        title: 'TreeSelect 分类选择',
+      },
+    ],
   },
   {
     groupName: '导航组件',
@@ -185,29 +193,29 @@ export default [
     list: [
       {
         path: '/grid',
-        title: 'Grid 宫格'
+        title: 'Grid 宫格',
       },
       {
         path: '/index-bar',
-        title: 'IndexBar 索引栏'
+        title: 'IndexBar 索引栏',
       },
       {
         path: '/sidebar',
-        title: 'Sidebar 侧边导航'
+        title: 'Sidebar 侧边导航',
       },
       {
         path: '/nav-bar',
-        title: 'NavBar 导航栏'
+        title: 'NavBar 导航栏',
       },
       {
         path: '/tab',
-        title: 'Tab 标签页'
+        title: 'Tab 标签页',
       },
       {
         path: '/tabbar',
-        title: 'Tabbar 标签栏'
-      }
-    ]
+        title: 'Tabbar 标签栏',
+      },
+    ],
   },
   {
     groupName: '业务组件',
@@ -215,20 +223,20 @@ export default [
     list: [
       {
         path: '/area',
-        title: 'Area 省市区选择'
+        title: 'Area 省市区选择',
       },
       {
         path: '/card',
-        title: 'Card 商品卡片'
+        title: 'Card 商品卡片',
       },
       {
         path: '/submit-bar',
-        title: 'SubmitBar 提交订单栏'
+        title: 'SubmitBar 提交订单栏',
       },
       {
         path: '/goods-action',
-        title: 'GoodsAction 商品导航'
-      }
-    ]
-  }
+        title: 'GoodsAction 商品导航',
+      },
+    ],
+  },
 ];