+++ packages/area/index.ts
@@ -1,6 +1,7 @@
 import { VantComponent } from '../common/component';
 import { pickerProps } from '../picker/shared';
 import { Weapp } from 'definitions/weapp';
+import { requestAnimationFrame } from '../common/utils';
 
 type AreaItem = {
   name: string;
@@ -56,9 +57,9 @@ VantComponent({
   },
 
   mounted() {
-    setTimeout(() => {
+    requestAnimationFrame(() => {
       this.setValues();
-    }, 0);
+    });
   },
 
   methods: {
@@ -201,31 +202,36 @@ VantComponent({
       }
 
       const stack = [];
+      const indexes = [];
+      const { columnsNum } = this.data;
 
-      stack.push(picker.setColumnValues(0, province, false));
-      stack.push(picker.setColumnValues(1, city, false));
-
-      if (city.length && code.slice(2, 4) === '00') {
-        [{ code }] = city;
+      if (columnsNum >= 1) {
+        stack.push(picker.setColumnValues(0, province, false));
+        indexes.push(this.getIndex('province', code));
       }
 
-      stack.push(
-        picker.setColumnValues(
-          2,
-          this.getList('county', code.slice(0, 4)),
-          false
-        )
-      );
+      if (columnsNum >= 2) {
+        stack.push(picker.setColumnValues(1, city, false));
+        indexes.push(this.getIndex('city', code));
+        if (city.length && code.slice(2, 4) === '00') {
+          [{ code }] = city;
+        }
+      }
+
+      if (columnsNum === 3) {
+        stack.push(
+          picker.setColumnValues(
+            2,
+            this.getList('county', code.slice(0, 4)),
+            false
+          )
+        );
+        indexes.push(this.getIndex('county', code));
+      }
 
       return Promise.all(stack)
         .catch(() => {})
-        .then(() =>
-          picker.setIndexes([
-            this.getIndex('province', code),
-            this.getIndex('city', code),
-            this.getIndex('county', code),
-          ])
-        )
+        .then(() => picker.setIndexes(indexes))
         .catch(() => {});
     },