+++ packages/cell/index.ts
@@ -31,6 +31,7 @@ VantComponent({
       type: Boolean,
       value: true,
     },
+    titleStyle: String,
   },
 
   methods: {