+++ packages/grid/index.ts
@@ -37,6 +37,14 @@ VantComponent({
       value: true,
       observer: 'updateChildren',
     },
+    direction: {
+      type: String,
+      observer: 'updateChildren',
+    },
+    iconSize: {
+      type: String,
+      observer: 'updateChildren',
+    },
   },
 
   data: {