+++ packages/mixins/transition.ts
@@ -1,4 +1,4 @@
-import { isObj } from '../common/utils';
+import { isObj, requestAnimationFrame } from '../common/utils';
 
 const getClassNames = (name: string) => ({
   enter: `van-${name}-enter van-${name}-enter-active enter-class enter-active-class`,
@@ -6,8 +6,6 @@ const getClassNames = (name: string) => ({
   leave: `van-${name}-leave van-${name}-leave-active leave-class leave-active-class`,
   'leave-to': `van-${name}-leave-to van-${name}-leave-active leave-to-class leave-active-class`,
 });
-
-const nextTick = () => new Promise((resolve) => setTimeout(resolve, 1000 / 30));
 
 export const transition = function (showDefaultValue: boolean) {
   return Behavior({
@@ -53,29 +51,24 @@ export const transition = function (showDefaultValue: boolean) {
         this.status = 'enter';
         this.$emit('before-enter');
 
-        Promise.resolve()
-          .then(nextTick)
-          .then(() => {
-            this.checkStatus('enter');
-            this.$emit('enter');
+        requestAnimationFrame(() => {
+          this.checkStatus('enter');
+          this.$emit('enter');
 
-            this.setData({
-              inited: true,
-              display: true,
-              classes: classNames.enter,
-              currentDuration,
-            });
-          })
-          .then(nextTick)
-          .then(() => {
+          this.setData({
+            inited: true,
+            display: true,
+            classes: classNames.enter,
+            currentDuration,
+          });
+
+          requestAnimationFrame(() => {
             this.checkStatus('enter');
             this.transitionEnded = false;
 
-            this.setData({
-              classes: classNames['enter-to'],
-            });
-          })
-          .catch(() => {});
+            this.setData({ classes: classNames['enter-to'] });
+          });
+        });
       },
 
       leave() {
@@ -90,28 +83,23 @@ export const transition = function (showDefaultValue: boolean) {
         this.status = 'leave';
         this.$emit('before-leave');
 
-        Promise.resolve()
-          .then(nextTick)
-          .then(() => {
-            this.checkStatus('leave');
-            this.$emit('leave');
+        requestAnimationFrame(() => {
+          this.checkStatus('leave');
+          this.$emit('leave');
 
-            this.setData({
-              classes: classNames.leave,
-              currentDuration,
-            });
-          })
-          .then(nextTick)
-          .then(() => {
+          this.setData({
+            classes: classNames.leave,
+            currentDuration,
+          });
+
+          requestAnimationFrame(() => {
             this.checkStatus('leave');
             this.transitionEnded = false;
             setTimeout(() => this.onTransitionEnd(), currentDuration);
 
-            this.setData({
-              classes: classNames['leave-to'],
-            });
-          })
-          .catch(() => {});
+            this.setData({ classes: classNames['leave-to'] });
+          });
+        });
       },
 
       checkStatus(status: 'enter' | 'leave') {