+++ packages/notice-bar/index.ts
@@ -1,8 +1,6 @@
 import { VantComponent } from '../common/component';
 import { Weapp } from 'definitions/weapp';
-
-const FONT_COLOR = '#ed6a0c';
-const BG_COLOR = '#fffbe8';
+import { requestAnimationFrame } from '../common/utils';
 
 VantComponent({
   props: {
@@ -48,14 +46,9 @@ VantComponent({
       type: String,
       value: '',
     },
-    color: {
-      type: String,
-      value: FONT_COLOR,
-    },
-    backgroundColor: {
-      type: String,
-      value: BG_COLOR,
-    },
+    color: String,
+    backgroundColor: String,
+    background: String,
     wrapable: Boolean,
   },
 
@@ -120,25 +113,28 @@ VantComponent({
           .export(),
       });
 
-      setTimeout(() => {
+      requestAnimationFrame(() => {
         this.setData({
           animationData: this.animation
             .translateX(-this.contentWidth)
             .step()
             .export(),
         });
-      }, 20);
+      });
 
       this.timer = setTimeout(() => {
         this.scroll();
       }, this.duration);
     },
 
-    onClickIcon() {
-      this.timer && clearTimeout(this.timer);
-      this.timer = null;
+    onClickIcon(event) {
+      if (this.data.mode === 'closeable') {
+        this.timer && clearTimeout(this.timer);
+        this.timer = null;
 
-      this.setData({ show: false });
+        this.setData({ show: false });
+        this.$emit('close', event.detail);
+      }
     },
 
     onClick(event: Weapp.Event) {