+++ packages/slider/index.ts
@@ -26,11 +26,15 @@ VantComponent({
     value: {
       type: Number,
       value: 0,
-      observer: 'updateValue',
+      observer(val) {
+        if (val !== this.value) {
+          this.updateValue(val);
+        }
+      },
     },
     barHeight: {
       type: null,
-      value: '2px',
+      value: 2,
     },
   },
 
@@ -43,7 +47,7 @@ VantComponent({
       if (this.data.disabled) return;
 
       this.touchStart(event);
-      this.startValue = this.format(this.data.value);
+      this.startValue = this.format(this.value);
       this.dragStatus = 'start';
     },
 
@@ -94,8 +98,9 @@ VantComponent({
       const { min } = this.data;
       const width = `${((value - min) * 100) / this.getRange()}%`;
 
+      this.value = value;
+
       this.setData({
-        value,
         barStyle: `
           width: ${width};
           ${drag ? 'transition: none;' : ''}