+++ packages/tabs/index.ts
@@ -26,36 +26,37 @@ VantComponent({
         }
       );
       this.updateTabs();
-    }
+    },
   },
 
   props: {
+    sticky: Boolean,
+    border: Boolean,
+    swipeable: Boolean,
+    titleActiveColor: String,
+    titleInactiveColor: String,
     color: {
       type: String,
-      observer: 'setLine'
+      observer: 'setLine',
     },
-    sticky: Boolean,
     animated: {
       type: Boolean,
       observer() {
         this.children.forEach((child: TrivialInstance, index: number) =>
           child.updateRender(index === this.data.currentIndex, this)
         );
-      }
+      },
     },
-    swipeable: Boolean,
     lineWidth: {
       type: [String, Number],
-      value: -1,
-      observer: 'setLine'
+      value: 40,
+      observer: 'setLine',
     },
     lineHeight: {
       type: [String, Number],
       value: -1,
-      observer: 'setLine'
+      observer: 'setLine',
     },
-    titleActiveColor: String,
-    titleInactiveColor: String,
     active: {
       type: [String, Number],
       value: 0,
@@ -63,45 +64,41 @@ VantComponent({
         if (name !== this.getCurrentName()) {
           this.setCurrentIndexByName(name);
         }
-      }
+      },
     },
     type: {
       type: String,
-      value: 'line'
-    },
-    border: {
-      type: Boolean,
-      value: true
+      value: 'line',
     },
     ellipsis: {
       type: Boolean,
-      value: true
+      value: true,
     },
     duration: {
       type: Number,
-      value: 0.3
+      value: 0.3,
     },
     zIndex: {
       type: Number,
-      value: 1
+      value: 1,
     },
     swipeThreshold: {
       type: Number,
-      value: 4,
+      value: 5,
       observer(value) {
         this.setData({
-          scrollable: this.children.length > value || !this.data.ellipsis
+          scrollable: this.children.length > value || !this.data.ellipsis,
         });
-      }
+      },
     },
     offsetTop: {
       type: Number,
-      value: 0
+      value: 0,
     },
     lazyRender: {
       type: Boolean,
-      value: true
-    }
+      value: true,
+    },
   },
 
   data: {
@@ -111,7 +108,7 @@ VantComponent({
     scrollable: false,
     trackStyle: '',
     currentIndex: null,
-    container: null
+    container: null,
   },
 
   mounted() {
@@ -124,7 +121,7 @@ VantComponent({
   methods: {
     updateContainer() {
       this.setData({
-        container: () => this.createSelectorQuery().select('.van-tabs')
+        container: () => this.createSelectorQuery().select('.van-tabs'),
       });
     },
 
@@ -132,7 +129,8 @@ VantComponent({
       const { children = [], data } = this;
       this.setData({
         tabs: children.map((child: TrivialInstance) => child.data),
-        scrollable: this.children.length > data.swipeThreshold || !data.ellipsis
+        scrollable:
+          this.children.length > data.swipeThreshold || !data.ellipsis,
       });
 
       this.setCurrentIndexByName(this.getCurrentName() || data.active);
@@ -150,7 +148,7 @@ VantComponent({
       this.$emit(eventName, {
         index: currentChild.index,
         name: currentChild.getComputedName(),
-        title: currentChild.data.title
+        title: currentChild.data.title,
       });
     },
 
@@ -235,7 +233,7 @@ VantComponent({
         duration,
         currentIndex,
         lineWidth,
-        lineHeight
+        lineHeight,
       } = this.data;
 
       this.getRect('.van-tab', true).then(
@@ -244,7 +242,6 @@ VantComponent({
           if (rect == null) {
             return;
           }
-          const width = lineWidth !== -1 ? lineWidth : rect.width / 2;
           const height =
             lineHeight !== -1
               ? `height: ${addUnit(lineHeight)}; border-radius: ${addUnit(
@@ -256,7 +253,7 @@ VantComponent({
             .slice(0, currentIndex)
             .reduce((prev, curr) => prev + curr.width, 0);
 
-          left += (rect.width - width) / 2;
+          left += (rect.width - lineWidth) / 2;
 
           const transition = skipTransition
             ? ''
@@ -265,12 +262,12 @@ VantComponent({
           this.setData({
             lineStyle: `
             ${height}
-            width: ${addUnit(width)};
+            width: ${addUnit(lineWidth)};
             background-color: ${color};
             -webkit-transform: translateX(${left}px);
             transform: translateX(${left}px);
             ${transition}
-          `
+          `,
           });
         }
       );
@@ -286,7 +283,7 @@ VantComponent({
 
       Promise.all([
         this.getRect('.van-tab', true),
-        this.getRect('.van-tabs__nav')
+        this.getRect('.van-tabs__nav'),
       ]).then(
         ([tabRects, navRect]: [
           WechatMiniprogram.BoundingClientRectCallbackResult[],
@@ -298,7 +295,7 @@ VantComponent({
             .reduce((prev, curr) => prev + curr.width, 0);
 
           this.setData({
-            scrollLeft: offsetLeft - (navRect.width - tabRect.width) / 2
+            scrollLeft: offsetLeft - (navRect.width - tabRect.width) / 2,
           });
         }
       );
@@ -324,17 +321,39 @@ VantComponent({
     onTouchEnd() {
       if (!this.data.swipeable) return;
 
-      const { tabs, currentIndex } = this.data;
       const { direction, deltaX, offsetX } = this;
       const minSwipeDistance = 50;
 
       if (direction === 'horizontal' && offsetX >= minSwipeDistance) {
-        if (deltaX > 0 && currentIndex !== 0) {
-          this.setCurrentIndex(currentIndex - 1);
-        } else if (deltaX < 0 && currentIndex !== tabs.length - 1) {
-          this.setCurrentIndex(currentIndex + 1);
+        const index = this.getAvaiableTab(deltaX);
+        if (index !== -1) {
+          this.setCurrentIndex(index);
         }
       }
-    }
-  }
+    },
+
+    getAvaiableTab(direction: number) {
+      const { tabs, currentIndex } = this.data;
+      const step = direction > 0 ? -1 : 1;
+
+      for (
+        let i = step;
+        currentIndex + i < tabs.length && currentIndex + i >= 0;
+        i += step
+      ) {
+        const index = currentIndex + i;
+
+        if (
+          index >= 0 &&
+          index < tabs.length &&
+          tabs[index] &&
+          !tabs[index].disabled
+        ) {
+          return index;
+        }
+      }
+
+      return -1;
+    },
+  },
 });