+++ packages/uploader/index.ts
@@ -12,7 +12,7 @@ VantComponent({
     beforeRead: null,
     previewSize: {
       type: null,
-      value: 90,
+      value: 80,
     },
     name: {
       type: [Number, String],
@@ -77,6 +77,8 @@ VantComponent({
           typeof item.isImage === 'undefined'
             ? isImageFile(item)
             : item.isImage,
+        deletable:
+          typeof item.deletable === 'undefined' ? true : item.deletable,
       }));
       this.setData({ lists, isInCount: lists.length < maxCount });
     },